backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic-settings
google-generativeai
sqlalchemy
psycopg2-binary
python-jose[cryptography]
google-auth
google-auth-oauthlib
requests
pydantic
alembic
